Added 'BLINDEZONE_GATES' argument
Controlling the number of gates in the output L2-product.
Due to various technical resasons, the first number of range gates might be not usuable. Because this number varies for different instruments, the decision wether to include or exclude the first N gates can now be specified in the configuration File.
E.g. 'BLINDEZONE_GATES=2' removes the first two range gates from the output L2-product.