Toolbox update 2022/10/30
Added functionality for windcube DBS and VAD retrieval. This should make the toolbox accessable to a wider range of users. Note that the adjustments to the code have been made on the basis of only a few windcube example files. Therefore, users are strongly encouraged to report any bugs or issues they encounter while using the client software.