Skipping corrupted netCDF (1/2)
Fixed an issue related to reading corrupted netCDF files. The software now skips over such files, but displays the skipped file and its directory. This issue also raised another problem related to backscatter not being present in the respective data. This made changes to the processing and reading utilities necessary.